requirements.txt
greenlet==3.1.1
inexactsearch==1.0.2
joblib==1.4.2
numpy==2.1.2
pandas==2.2.3
PyMySQL==1.1.1
pyspellchecker==0.8.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.2.0
silpa_common==0.3
six==1.16.0
soundex==1.1.3
SQLAlchemy==2.0.31
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
